requirements.txt
# SenorScout - AI-Powered Price Scanner
# Core dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP & Scraping
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
urllib3>=2.2.0

# ML & Computer Vision
ultralytics>=8.1.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.2.0
opencv-python>=4.9.0

# Data Processing
numpy>=1.26.0
pandas>=2.2.0

# Validation & Settings
pydantic>=2.5.0
python-dotenv>=1.0.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.1
